% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advance_population.R
\name{advance_population}
\alias{advance_population}
\title{Advances the population by one year.}
\usage{
advance_population(.data)
}
\arguments{
\item{.data}{data frame including the starting population of each
demographic group; either the initial population or the previous projected
year. Possible identifiers are the same as in \code{parameters} (apart from year).
The data frame only includes one year, usually the one preceding the first
projected year.
\itemize{
\item \code{year} numeric
\item \code{scen}, character, one or several projection scenario(s). The main
scenarios are usually "reference", "low" growth, and "high" growth.
\item \code{spatial_unit} character.
\item \code{nat} character.
\item \code{sex} character.
\item \code{age} numeric.
\item \code{n_dec} numeric, number of people per demographic group in December.
}}
}
\value{
returns an identically structured data frame with the aged population.
}
\description{
Advances the population by one year; increases age by one year and aggregates
people aged 100 and older.
}
