% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_migration_subregional.R
\docType{data}
\name{ag_migration_subregional}
\alias{ag_migration_subregional}
\title{Subregional migration records for the canton of Aargau (five subregions)}
\format{
The example data includes historic migration records for immigration and
emigration. Dimensions correspond to the number of people of each
demographic group (nationality (2) X sex (2) X age classes (101)) for five
spatial units of the canton of Aargau in 2025.
}
\usage{
ag_migration_subregional
}
\description{
Raw migration data from the canton of Aargau divided into five subregions.
}
\keyword{datasets}
