% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_population_subregional.R
\docType{data}
\name{ag_population_subregional}
\alias{ag_population_subregional}
\title{Population data for the canton of Aargau (five subregions)}
\format{
The example population records include the number of people of each
demographic group (nationality (2) X sex (2) X age classes (101)) for five
spatial units of the canton of Aargau in 2025.
}
\usage{
ag_population_subregional
}
\description{
Population from the canton of Aargau divided into five subregions.
}
\keyword{datasets}
