% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_measures.R
\name{aggregate_measures}
\alias{aggregate_measures}
\title{Aggregate evaluation measures}
\usage{
aggregate_measures(data, weight_groups = NULL)
}
\arguments{
\item{data}{data frame created with function \code{compute_measures}.}

\item{weight_groups}{character, optional argument indicating one or more
column names to obtain evaluation criteria weighted for specific groups
(e.g., age groups, nationality).}
}
\value{
#' A data frame. The data frame includes the following summary measures:
\itemize{
\item \code{mpe} is the mean percentage error
(\bold{mpe}; or mean algebraic percentage error
\bold{malpe}); it is a bias indicator as it takes the \bold{direction} of the
error into account. Positive values indicate that the projections were,
overall, too high. Negative values indicate that the projections were,
overall, too low.
The closer the value is too zero, the lower the bias.
\item \code{medpe} is the median (or middle value) of the percentage error
(\bold{medpe}). Particularly useful for small samples or skewed
distributions. The closer the value is too zero, the lower the bias.
\item \code{mape} is the mean \bold{absolute} percentage / proportional error
(\bold{mape}). It considers variance (or amplitude) and can be seen as a
measure of precision.
The smaller the value, the lower is the average error.
\item \code{medape} is the median (or middle value) of the \bold{absolute}
percentage error (\bold{medape}). Particularly useful for small samples or
skewed distributions. The smaller the value, the lower is the average error.
\item \code{rmse} is the root mean square error; it is an indication of the robustness
or quality of the projection. The smaller the value, the more robust the
projection.
\item \code{wmape} is the \bold{weighted} mean \bold{absolute} percentage
error (\bold{wmape}); in contrast to \code{mape}, this measure
weights each absolute percentage error according to the population size of
the "focal" group (e.g., nationality, age group) and thus considers domain
size. Put differently, errors count more in populous groups than in less
populous groups. This measure is particularly useful when population sizes
vary strongly. The closer the value, the more precise is the
projection.
\item \code{n_measure} is the number of times a summary measure occurs (per weight
group if requested).
\item \code{ape_under_1} is a measure to gauge the error distribution; it indicates the
proportion of observations that have absolute percentage errors smaller than
1\%.
\item \code{ape_under_5} is a measure to gauge the error distribution; it indicates the
proportion of observations that have absolute percentage errors smaller than
5\%.
}
}
\description{
Returns descriptive summary statistics of model accuracy and bias measures
across demographic groups and years. The returned statistics are particularly
useful for comparing the model performance for different groups or different
models.
}
\examples{
\dontrun{
# Get evaluation measures without weights
aggregate_measures(evaluation)
# Get evaluation measures weighted for groups
compute_measures(evaluation, weight_groups = "age")
}
}
\references{
Baker, J., et al. (2015). Sub-county population estimates using administrative records: A municipal-level case study in New Mexico. In M. N. Hoque & L. B. Potter (Eds.), Emerging techniques in applied demography (pp. 63-79). Springer, \url{https://doi.org/10.1007/978-94-017-8990-5_6}

Bérard-Chagnon, J. (2015) Using tax data to estimate the number of families and households in Canada. In M. N. Hoque & L. B. Potter (Eds.), Emerging techniques in applied demography (pp. 137-153). Springer, \url{https://doi.org/10.1007/978-94-017-8990-5_10}

Reinhold M. & Thomsen, S. L. (2015) Subnational population projections by age: An evaluation of combined forecast techniques, Population Research and Policy Review, 34, 593-613, \url{https://doi.org/10.1007/s11113-015-9362-0}

Wilson, T. (2012). Forecast accuracy and uncertainty of Australian Bureau of Statistics state and territory population projections, International Journal of Population Research, 1, 419824, \url{https://doi.org/10.1155/2012/419824}

Wilson, T. (2016). Evaluation of alternative cohort-component models for local area population forecasts, Population Research and Policy Review, 35, 241-261, \url{https://doi.org/10.1007/s11113-015-9380-y}
}
