% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_emi_mean.R
\name{calculate_emi_mean}
\alias{calculate_emi_mean}
\title{Get mean historical subregional emigration to municipalities within the canton}
\usage{
calculate_emi_mean(
  past_migration,
  n_jan,
  births,
  emi_n,
  spatial_unit,
  year_n = NULL
)
}
\arguments{
\item{past_migration}{data frame, containing aggregated migration records.
The data frame must include one row for each combination of year,
spatial unit, nationality, sex, and age.}

\item{n_jan}{character, column containing the starting population (typically
the number of people per group in January).}

\item{births}{character, column containing the number of births in the group
of 0-year olds.}

\item{emi_n}{character, column containing the total number of people emigrating per
spatial unit and demographic group.}

\item{spatial_unit}{character, column containing the spatial units.}

\item{year_n}{numeric, number of years to be considered for computing the mean.}
}
\value{
A data frame that includes the average emigration rate.
}
\description{
The function calculates the mean of past emigration to other
municipalities within the canton for all municipalities and all possible
combinations of age group, sex, and nationality.
}
\examples{
# Compute mean emigration rate
calculate_emi_mean(
past_migration = ag_migration_subregional,
n_jan = n_jan,
births = births,
emi_n = emi_n,
spatial_unit = spatial_unit,
year_n = NULL
)
}
