% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_newborns.R
\name{calculate_newborns}
\alias{calculate_newborns}
\title{Calculate newborns per demographic group.}
\usage{
calculate_newborns(
  population,
  parameters,
  fert_first,
  fert_last,
  share_born_female,
  subregional = FALSE
)
}
\arguments{
\item{population}{data frame including the starting population of each
demographic group; either the initial population or the previous
projected year.}

\item{parameters}{data frame containing the FSO rates and numbers to run the
projection for a specific spatial level (e.g., canton, municipality).}

\item{fert_first}{numeric, first year of female fertility. Defaults to 16
(FSO standard value).}

\item{fert_last}{numeric, last year of female fertility. Defaults to 50
(FSO standard value).}

\item{share_born_female}{numeric, fraction of female babies. Defaults to
100 / 205 (FSO standard value).}
}
\value{
Returns a data frame with the number of newborn children 'births'
structured in demographic groups.
}
\description{
Calculate newborns per demographic group.
}
