% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_projection.R
\name{calculate_projection}
\alias{calculate_projection}
\title{Calculates the projection for ages 1-100.}
\usage{
calculate_projection(.data, subregional = subregional)
}
\arguments{
\item{.data}{data frame, population and parameters for one year.}

\item{subregional}{character or NULL, indicates if subregional migration
patterns (e.g., movement between municipalities within a canton) are
part of the projection (default \code{subregional = NULL}). Requires input
on the level of subregions (in \code{parameters} and \code{population}).
Two calculation methods are supported to distribute people between
subregions: With \code{subregional = "net"}, the net migration between
subregions is added to the population balance. Net migration numbers
must be specified in a data column \code{mig_sub} in \code{parameters}.
With \code{subregional = "rate"}, the numbers for subregional emigrants are
subtracted from the population balance, then redistributed back to all
subregional units as subregional immigration; \code{parameters} must contain
the columns \code{emi_sub} and \code{imm_sub}.}
}
\value{
Returns a data frame with the starting population \code{n_jan}, components,
and the resulting projected population for the next year \code{n_dec}.
}
\description{
Uses the cohort component method to project population development.
}
