% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_rates.R
\name{calculate_rates}
\alias{calculate_rates}
\title{Calculate emigration rates to subregions}
\usage{
calculate_rates(
  past_migration,
  n_jan,
  births,
  emi_n,
  spatial_unit,
  method,
  year_range = NULL,
  age_group = 1,
  binational = TRUE,
  two_sex = TRUE
)
}
\arguments{
\item{past_migration}{data frame, containing aggregated migration records.
The data frame must include one row for each combination of year,
spatial unit, nationality, sex, and age.}

\item{n_jan}{character, column containing the starting population (typically
the number of people per group in January).}

\item{births}{character, column containing the number of births in the group
of 0-year olds.}

\item{emi_n}{character, column containing the total number of people emigrating per
spatial unit and demographic group.}

\item{spatial_unit}{character, column containing the spatial units.}

\item{method}{character, method to calculate average shares, i.e. \code{mean} or
\code{median}.}

\item{year_range}{\bold{(optional)} vector, years taken into consideration to
calculate historical shares. Default uses all years present in the data.}

\item{age_group}{\bold{(optional)} integer, divides continuous age values into
intervals for calculating shares. If the argument is not specified,
the default uses 1-year age groups.}

\item{binational}{\bold{(optional)} boolean, \code{TRUE} indicates that the calculation
discriminates between two groups of nationalities. \code{FALSE} indicates
that the calculation does not distinguish between nationalities.}

\item{two_sex}{\bold{(optional)} boolean, \code{TRUE} indicates that the calculation
discriminates between two sexes. \code{FALSE} indicates that the calculation
does not distinguish between sexes.}
}
\value{
A data frame that includes the average emigration rate.
}
\description{
The function calculates the mean of past emigration to other
municipalities within the canton for all municipalities and all possible
combinations of age group, sex, and nationality.
}
\examples{
# Compute mean emigration rate
calculate_rates(
 past_migration = ag_migration_subregional,
 n_jan = n_jan,
 births = births,
 emi_n = hist_emi_sub_n,
 spatial_unit = spatial_unit,
 method = "mean",
 year_range = c(2022:2024),
 age_group = 5,
 binational = TRUE,
 two_sex = TRUE
)
}
