% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_shares.R
\name{calculate_shares}
\alias{calculate_shares}
\title{Calculate shares for distributing immigration among subregions}
\usage{
calculate_shares(
  past_migration,
  imm_n,
  year_range = NULL,
  age_group = 1,
  binational = TRUE,
  two_sex = TRUE
)
}
\arguments{
\item{past_migration}{data frame, historical records (e.g., immigration from
other cantons, countries or subregions). Required columns are: \code{year},
\code{spatial_unit}, \code{age} and a column that contains aggregated historical
migration records. The columns \code{nat} and \code{sex} are optional.}

\item{imm_n}{character, name of the column which contains the
data for aggregated historical migration records.}

\item{year_range}{\bold{(optional)} vector, years taken into consideration to
calculate historical shares. Default uses all years present in the data.}

\item{age_group}{\bold{(optional)} integer, divides continuous age values into
intervals for calculating shares. If the argument is not specified,
the default uses 1-year age groups.}

\item{binational}{\bold{(optional)} boolean, \code{TRUE} indicates that the calculation
discriminates between two groups of nationalities. \code{FALSE} indicates
that the calculation does not distinguish between nationalities.}

\item{two_sex}{\bold{(optional)} boolean, \code{TRUE} indicates that the calculation
discriminates between two sexes. \code{FALSE} indicates that the calculation
does not distinguish between sexes.}
}
\value{
A data frame that includes the average share per demographic group and spatial
unit. \code{imm_share} can be used as \code{imm_sub} parameter when \code{propop::propop()} uses
\code{subregional = "rate"}.
}
\description{
Calculates historical immigration shares across spatial units
within a region. \cr
To compute the shares, the counts are summed within
each spatial unit and across all spatial units combined, across all
years. Dividing the spatial unit count by the total gives the group's
share.\cr
These shares are used to allocate emigrants moving from one
subregion to another subregion (which can be done with \code{calculate_rates}).
}
\examples{
# Calculate shares to distribute subregional immigration among spatial units
calculate_shares(
  past_migration = ag_migration_subregional,
  imm_n = hist_imm_sub_n,
  year_range = c(2022:2024),
  age_group = 10,
  binational = TRUE,
  two_sex = TRUE
)
}
\seealso{
\code{\link[=propop]{propop()}} for details on how to account for subregional migration using the rate method,
\code{\link[=calculate_rates]{calculate_rates()}} for calculating the associated emigration rate \code{emi_sub}.
}
