% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_balance.R
\name{check_balance}
\alias{check_balance}
\title{Check if population equation components add up within one year}
\usage{
check_balance(data)
}
\arguments{
\item{data}{data frame containing population projections; can be
created with \code{propop::propop()}.}
}
\value{
data frame containing summary statistics.
}
\description{
Function to check integrity of components in \code{propop::propop()} output.
Takes population at the beginning of the year (\code{n_jan}), adds all components
(births - mor - emi_int - emi_nat + imm_int + imm_nat + acq) and checks if
the sum is equal to the population at the end of the year (\code{n_dec}).
}
\examples{
\dontrun{
projection_canton_2030 <- propop(
parameters = fso_parameters,
year_first = 2025,
year_last = 2035,
population = fso_population,
subregional = FALSE,
binational = TRUE
)
check_balance(projection_canton_2030)
}
}
