% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fso_parameters.R
\docType{data}
\name{fso_parameters}
\alias{fso_parameters}
\title{Sample parameters to run population projection}
\format{
The example data include the required parameters for each demographic group
(nationality (2) X sex (2) X age classes (101)) for the years 2019-2030.
}
\source{
Data obtained from the Swiss Federal Statistical Office (FSO):
\itemize{
\item \url{https://www.pxweb.bfs.admin.ch/pxweb/en/px-x-0104020000_101/-/px-x-0104020000_101.px/}
\item \url{https://www.pxweb.bfs.admin.ch/pxweb/en/px-x-0104020000_102/-/px-x-0104020000_102.px}
\item \url{https://www.pxweb.bfs.admin.ch/pxweb/en/px-x-0104020000_103/-/px-x-0104020000_103.px/}
\item \url{https://www.pxweb.bfs.admin.ch/pxweb/en/px-x-0104020000_106/-/px-x-0104020000_106.px/}
\item \url{https://www.pxweb.bfs.admin.ch/pxweb/en/px-x-0104020000_109/-/px-x-0104020000_109.px/}
}
}
\usage{
fso_parameters
}
\description{
Data frame containing the rates and number of people from the
\href{https://www.bfs.admin.ch/bfs/en/home.html}{Federal Statistical Office (FSO)}
required to project the development of four demographic groups for a selected
canton (Aargau). The parameters are from the
\href{https://www.bfs.admin.ch/bfs/en/home/statistics/population/population-projections/national-projections.html}{model published in 2020}.
The sample data only include the reference scenario and the years 2019-2030.
}
\section{Demographic groups}{

The returned data frame includes parameters for each unique combination of
the following demographic variables:
\itemize{
\item \code{nat}: ch = Swiss; int = foreign / international.
\item \code{sex}: f = female, m = male.
\item \code{age}: 101 one-year age classes, ranging from 0 to 100 (including those
older than 100).
}
}

\section{Parameters}{

The following parameters are included in the returned data frame:
\itemize{
\item \code{year}: character, year of projection.
\item \code{scen}: character, projection scenario.
\item \code{birth_rate}: numeric, number of children per year.
\item \code{births_int_ch}: numeric, proportion of children with Swiss nationality
born to non-Swiss mothers.
\item \code{mor}: numeric, prospective mortality (probability of death).
\item \code{emi}: numeric, rate of people emigrating abroad.
\item \code{acq}: numeric, rate of acquisition of Swiss citizenship.
\item \code{imm_int}: numeric, number of people immigrating from abroad.
\item \code{mig_ch}: numeric, national / inter-cantonal net migration
(number of immigrants minus number of emigrants).
\item \code{spatial_unit}: character, indicating the user requested spatial
unit(s).
}
}

\section{Details about calculated variables}{


\code{births_int_ch} is calculated by dividing the number of live newborns with
Swiss citizenship born to non-Swiss mothers by the number of all live
newborns born to non-Swiss mothers.

\code{mig_ch} is calculated as the difference between the immigration from other
cantons and the emigration to other cantons.
}

\keyword{datasets}
