% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fso_projections.R
\docType{data}
\name{fso_projections}
\alias{fso_projections}
\title{Sample population projections from FSO for the canton of Aargau}
\format{
The example projections include the projected number of people of each
demographic group (nationality (2) X sex (2) X age classes (101)) for the
years 2019-2030.
}
\source{
Federal Statistical Office: \url{https://www.pxweb.bfs.admin.ch/pxweb/en/px-x-0104020000_101/-/px-x-0104020000_101.px/}
}
\usage{
fso_projections
}
\description{
Data frame containing the population projections from the
Federal Statistical Office (FSO) for four demographic groups for the canton
of Aargau. The projections are from the model published in 2020. The sample
data only include the reference scenario and the years 2019-2030.
}
\section{Variables}{

\itemize{
\item \code{year}: character, year of projection.
\item \code{scen}: character, only including the "reference" scenario.
\item \code{nat}: character, ch = Swiss; int = foreign / international.
\item \code{sex}: character, f = female, m = male.
\item \code{age}: numeric, 101 one-year age classes, ranging from 0 to 100
(including those older than 100).
\item \code{spatial_unit}: character, indicating that the data refer to the canton
Aargau.
\item \code{n_projected}: numeric, number of people per demographic group and year.
}
}

\keyword{datasets}
