% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters.R
\name{get_parameters}
\alias{get_parameters}
\title{Get projection parameters from FSO}
\source{
Data obtained from the Swiss Federal Statistical Office (FSO):
\itemize{
\item \url{https://www.pxweb.bfs.admin.ch/pxweb/en/px-x-0104020000_101/-/px-x-0104020000_101.px/}
\item \url{https://www.pxweb.bfs.admin.ch/pxweb/en/px-x-0104020000_102/-/px-x-0104020000_102.px}
\item \url{https://www.pxweb.bfs.admin.ch/pxweb/en/px-x-0104020000_103/-/px-x-0104020000_103.px/}
\item \url{https://www.pxweb.bfs.admin.ch/pxweb/en/px-x-0104020000_106/-/px-x-0104020000_106.px/}
\item \url{https://www.pxweb.bfs.admin.ch/pxweb/en/px-x-0104020000_109/-/px-x-0104020000_109.px/}
}
}
\usage{
get_parameters(
  number_fso_ref = "px-x-0104020000_101",
  number_fso_high = "px-x-0104020000_102",
  number_fso_low = "px-x-0104020000_103",
  number_fso_rates = "px-x-0104020000_109",
  number_fso_births = "px-x-0104020000_106",
  year_first,
  year_last,
  spatial_units
)
}
\arguments{
\item{number_fso_ref}{character, px-x table ID for number parameters
(reference scenario), defaults to "px-x-0104020000_101".}

\item{number_fso_high}{character, px-x table ID for number parameters
(high growth scenario), defaults to "px-x-0104020000_102".}

\item{number_fso_low}{character, px-x table ID for for number parameters
(low growth scenario, defaults to "px-x-0104020000_103".}

\item{number_fso_rates}{character, px-x table ID for rate parameters,
defaults to "px-x-0104020000_109".}

\item{number_fso_births}{character, px-x table ID required to compute the
share of Swiss newborns from non-Swiss mothers,
defaults to "px-x-0104020000_106".}

\item{year_first}{numeric,  first year for which the parameters and
projections are to be downloaded.}

\item{year_last}{numeric, last year for which the parameters and projections
are to be downloaded.}

\item{spatial_units}{character vector, indicating at least one spatial
entity for which the projection will be run. Typically a canton.}
}
\value{
A data frame with the rates and number of people from the
\href{https://www.bfs.admin.ch/bfs/en/home.html}{Federal Statistical Office (FSO)}
required to project the population development of the requested spatial
entities. For each of the four demographic groups (nationality x sex),
there are 101 age classes, resulting in a total of 404 rows per requested
year and spatial unit.
}
\description{
Users who do not have the mandatory projection parameters for
\code{propop::propop()} can use this convenience function to download them from
the Federal Statistical Office (FSO). The parameters are only available
on the level of cantons. For smaller-scale projections, the parameters must
be scaled down.
In addition to the parameters, the function also returns the projected
population (i.e., number of people estimated in the FSO model released in 2020).
All parameters and projections are from the
\href{https://www.bfs.admin.ch/bfs/en/home/statistics/population/population-projections/national-projections.html}{FSO model published in 2020}.
The variables \code{int_mothers} and \code{mig_nat_n} are not directly available from
the FSO. They are calculated within this function.

To get projection parameters, you must use the spelling defined in the
corresponding FSO table. See
\code{vignette("prepare_data", package = "propop")}.

Changes to the API interface may break this function. If problems occur,
we recommend following the step-by-step procedure described in
\code{vignette("prepare_data", package = "propop")}.
}
\section{Demographic groups}{

The returned data frame includes parameters for each unique combination of
the following demographic variables:
\itemize{
\item \code{nat}: ch = Swiss; int = foreign / international.
\item \code{sex}: f = female, m = male.
\item \code{age}: 101 one-year age classes, ranging from 0 to 100 (including those
older than 100).
}
}

\section{Parameters}{

The following parameters are included in the returned data frame:
\itemize{
\item \code{year}: numeric, year of projection.
\item \code{scen}: character, projection scenario.
\item \code{birthrate}: numeric, total number of live human births per 1,000
inhabitants.
(formerly \code{birth_rate}).
\item \code{int_mothers}: numeric, proportion of children with Swiss nationality
born to non-Swiss mothers
(formerly \code{births_int_ch}).
\item \code{mor}: numeric, prospective mortality (probability of death).
\item \code{emi_int}: numeric, rate of people emigrating to other countries
(formerly \code{emi}).
\item \code{emi_nat}: numeric, rate of people emigrating to other cantons
(new parameter).
\item \code{acq}: numeric, rate of acquisition of Swiss citizenship.
\item \code{imm_int_n}: numeric, number of people immigrating from abroad
(formelry \code{imm_int}).
\item \code{imm_nat_n}: numeric, number of people immigrating from other cantons
(new parameter).
\item \code{emi_nat_n}: numeric, number of people emigrating to other cantons
(parameter previously used to compute \code{mig_nat_n}).
\item \code{mig_nat_n}: numeric, national / inter-cantonal net migration
(number of immigrants minus number of emigrants).
(formerly \code{mig_ch}, will soon be obsolete and removed).
\item \code{spatial_unit}: character, indicating the user requested spatial
unit(s).
}
}

\section{Projected population}{

\code{n_projected} is the the number of people per demographic group and year on
December 31 (as projected by the FSO in the 2020 model).
}

\section{Details about calculated variables}{


\code{births_int_ch} is calculated by dividing the number of live newborns with
Swiss citizenship born to non-Swiss mothers by the number of all live
newborns born to non-Swiss mothers.

\code{mig_ch} is calculated as the difference between the immigration from other
cantons and the emigration to other cantons.
}

\examples{
\dontshow{if ((Sys.getenv("RUN_EXPENSIVE_TESTS") == "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
one_canton <- get_parameters(
  year_first = 2025,
  year_last = 2050,
  spatial_units = "Aargau"
)
two_cantons_4years <- get_parameters(
  year_first = 2018,
  year_last = 2021,
  spatial_units = c("Aargau", "Zug")
)
}
\dontshow{\}) # examplesIf}
}
