% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_population.R
\name{get_population}
\alias{get_population}
\title{Get population data from FSO}
\source{
Federal Statistical Office:
\url{https://www.pxweb.bfs.admin.ch/pxweb/en/px-x-0102010000_101/-/px-x-0102010000_101.px/}
}
\usage{
get_population(
  number_fso = "px-x-0102010000_101",
  year_first,
  year_last,
  spatial_units
)
}
\arguments{
\item{number_fso}{character, px-x table ID for population records,
defaults to \verb{px-x-0102010000_101}.}

\item{year_first}{numeric, first year for which the population records are to
be downloaded.}

\item{year_last}{numeric, last year for which the population records are to
be downloaded. When downloading the starting population for the
projection, this will be the same as \code{year_first}.

\code{year_first} when requesting the starting population for \code{propop::propop()}}

\item{spatial_units}{character vector, indicating at least one spatial
entity for which the projection will be run. Typically a canton, districts,
or municipalities.}
}
\value{
A data frame. For each of the four demographic groups (female / male,
Swiss / foreign nationals), there are 101 age classes, resulting in a total
of 404 rows per requested year and spatial unit. Columns included in the
returned data frame:
\describe{
\item{year}{character, year in which the population was recorded.}
\item{spatial_unit}{character, indicating the spatial entities (e.g.,
cantons, districts, municipalities).}
\item{nat}{character, ch = Swiss, int = foreign / international.}
\item{sex}{character f = female, m = male.}
\item{age}{numeric, 101 one-year age classes, ranging from 0 to 100
(including those older than 100).}
\item{n}{numeric, number of people per year, spatial entity, and
demographic group.}
}
}
\description{
The function can be used to get the mandatory starting population for
\code{propop::propop()} from the Federal Statistical Office (FSO). The function
can also be used to obtain historical population records (e.g., for model
performance evaluations).

To get population data, you must use the spelling defined in the
corresponding FSO table. For more details see
\code{vignette("prepare_data", package = "propop")}.

Changes to the API interface may break this function. If problems occur, see
\code{vignette("prepare_data", package = "propop")}.
}
\examples{
\dontshow{if ((Sys.getenv("RUN_EXPENSIVE_TESTS") == "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
get_population(
  number_fso = "px-x-0102010000_101",
  year_first = 2018,
  year_last = 2019,
  spatial_units = "- Aargau"
)
get_population(
  year_first = 2018,
  year_last = 2018,
  spatial_units = c("- Aargau", "......0301 Aarberg")
)
}
\dontshow{\}) # examplesIf}
}
