% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_evaluation.R
\name{prepare_evaluation}
\alias{prepare_evaluation}
\title{Prepare data for evaluation}
\usage{
prepare_evaluation(
  data_benchmark,
  data_projected,
  drop_start_year = FALSE,
  age_groups = NULL
)
}
\arguments{
\item{data_benchmark}{data frame containing benchmark data (e.g., actual /
official population records obtained with \code{propop::get_population()}).}

\item{data_projected}{data frame containing population projections; can be
created with \code{propop::propop()}.}

\item{drop_start_year}{logical, indicating if starting population shall be
removed from \code{data_projected}.}

\item{age_groups}{character, optional argument (\code{"age_groups_3"}) indicating
if the data shall be aggregated into the predefined three age groups
(0-19, 20-64, over 65 years). Using aggregated groups will lead to smaller
projection errors than using 101 age classes. Currently only one option is
available for aggregating age groups. Defaults to using 101 one-year age
classes.}
}
\value{
Returns a data frame with the number of people from the benchmark and
from the projection. Each row contains a unique combination of year,  spatial
unit, and demographic group.
}
\description{
This functions takes benchmark data (typically population records) and
population projections and prepares a combined data frame to evaluate the
performance of the projection.
For more details on usage, see
\code{vignette("evaluate", package = "propop")}.
}
\section{Input data and variables}{

Both input data frames must contain the following variables for the
\bold{same range of years}:
\describe{
\item{year}{character, year in which the population was recorded.}
\item{spatial_unit}{character, indicating the spatial entities (e.g.,
cantons, districts, municipalities).}
\item{nat}{character, ch = Swiss, int = foreign / international.}
\item{sex}{character, f = female, m = male.}
\item{age}{numeric, 101 one-year age classes, ranging from 0 to 100
(including those older than 100).}
\item{n}{numeric, number of people per year, spatial entity, and
demographic group.}
}
}

\examples{
\dontrun{
combined <- prepare_evaluation(
  data_benchmark = output_get_population,
  data_projected = output_propop,
  drop_start_year = TRUE
)
combined_grouped <- prepare_evaluation(
  data_benchmark = output_get_population,
  data_projected = output_propop,
  drop_start_year = TRUE,
  age_groups = "age_groups_3"
)
}
}
