% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_population.R
\name{project_population}
\alias{project_population}
\title{Project population development}
\usage{
project_population(
  population,
  parameters,
  fert_first = 16,
  fert_last = 50,
  share_born_female = 100/205,
  subregional = NULL,
  binational = TRUE
)
}
\arguments{
\item{population}{data frame including the starting population of each
demographic group; either the initial population or the previous projected
year. Possible identifiers are the same as in \code{parameters} (apart from year).
The data frame only includes one year, usually the one preceding the first
projected year.}

\item{parameters}{list, data frames containing the FSO rates and numbers to
run the projection for a specific spatial level (e.g., canton, municipality).}

\item{fert_first}{numeric, first year of female fertility. Defaults to 16
(FSO standard value).}

\item{fert_last}{numeric, last year of female fertility. Defaults to 50
(FSO standard value).}

\item{share_born_female}{numeric, fraction of female babies. Defaults to
100 / 205 (FSO standard value).}

\item{subregional}{character or NULL, indicates if subregional migration
patterns (e.g., movement between municipalities within a canton) are
part of the projection (default \code{subregional = NULL}). Requires input
on the level of subregions (in \code{parameters} and \code{population}).
Two calculation methods are supported to distribute people between
subregions: With \code{subregional = "net"}, the net migration between
subregions is added to the population balance. Net migration numbers
must be specified in a data column \code{mig_sub} in \code{parameters}.
With \code{subregional = "rate"}, the numbers for subregional emigrants are
subtracted from the population balance, then redistributed back to all
subregional units as subregional immigration; \code{parameters} must contain
the columns \code{emi_sub} and \code{imm_sub}.}

\item{binational}{boolean, \code{TRUE} indicates that projections discriminate
between two groups of nationalities. \code{FALSE} indicates that the
projection is run without distinguishing between nationalities.}
}
\value{
Returns a data frame that includes the number of people for each demographic
group per year (for projected years) and spatial unit. The number of
rows is the product of all years times all demographic groups times
all spatial units.
The output includes several \bold{identifiers} that indicate to which
demographic group, year, and spatial unit the results in the rows refer
to:
\item{year}{integer, indicating the projected years.}
\item{spatial_unit}{factor, spatial units for which the projection
was run (e.g., canton, districts, municipalities).}
\item{age}{integer, ranging from \code{0}n to \code{100} (including those older
than 100).}
\item{sex}{factor, female (f) and male (m).}
\item{nat}{factor, indicates if the nationality is Swiss (ch) or
international / foreign (int). This variable is only returned if
\code{binational = TRUE}.}
The output also includes columns related to the \bold{size and change
of the population:}
\item{n_jan}{numeric, start-of-year population per demographic group.}
\item{n_dec}{numeric, end-of-year population per demographic group.}
The \bold{components} that are used to project the development of the
population are also included in the output:
\item{births}{numeric, number of births (non-zero values are only
available for age = 0).}
\item{mor_n}{numeric, number of deaths.}
\item{emi_int}{numeric, number of people who emigrate
to other countries.}
\item{emi_nat_n}{numeric, number of people who emigrate
to other cantons.}
\item{imm_int_n}{numeric, number of people who immigrate
from other countries.}
\item{imm_nat_n}{numeric, number of people who immigrate
from other cantons.}
\item{acq_n}{numeric, number of people who acquire Swiss citizenship.}
}
\description{
This function is applied within the wrapper function \code{propop_tables()}
to iterate across years and spatial units.

The calculations involve three helper functions:
\itemize{
\item \code{advance_population()}: Advances the population by one year (increases
age by one year and aggregates people aged 100 and older).
\item \code{calculate_projection()}: Calculates the projection for ages 1-100.
\item \code{calculate_newborns()}: Calculate newborns per demographic group.
}
}
