% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_raw.R
\name{project_raw}
\alias{project_raw}
\title{Project population development (raw results)}
\usage{
project_raw(
  parameters,
  year_first,
  year_last,
  age_groups = 101,
  fert_first = 16,
  fert_last = 50,
  share_born_female = 100/205,
  n,
  subregional
)
}
\arguments{
\item{parameters}{data frame containing the FSO rates and numbers to run the
projection for a specific spatial level (e.g., canton, municipality).
\itemize{
\item \code{year}: projection year.
\item \code{spatial_unit}: ID of spatial entity (e.g., canton, municipality) for
which to run the projections.
\item \code{scen}: projection scenario, used to subset data frames with multiple
scenarios (r = reference, l = low growth, h = high growth scenario).
\item \code{nat}: nationality (ch = Swiss; int = foreign / international).
\item \code{sex}: sex (f = female, m = male).
\item \code{age}: age classes; typically ranging from 0 to 100 (incl. >100).
\item \code{birth_rate}: number of children per year.
\item \code{births_int_ch} proportion of children with Swiss nationality born to
non-Swiss mothers.
\item \code{mor}: prospective mortality rate (probability of death).
\item \code{acq}: rate of acquisition of Swiss citizenship.
\item \code{emi}: rate of people emigrating abroad.
\item \code{mig_ch}: national / inter-cantonal net migration
(number of immigrants - number of emigrants).
\item \code{imm_int}: number of people immigrating from abroad.
\item \code{mig_sub}: within canton net migration. Useful to account for movements
between different subregions (e.g., municipalities). This argument is
\bold{optional.}
}}

\item{year_first}{numeric, first year to be projected.}

\item{year_last}{numeric, last year to be projected.}

\item{age_groups}{numeric, number of age classes. Creates a vector with
1-year age classes running from \code{0} to (\code{age_groups} - 1). Defaults to
\code{101} (FSO standard number of age groups).}

\item{fert_first}{numeric, first year of female fertility. Defaults to 16
(FSO standard value).}

\item{fert_last}{numeric, last year of female fertility. Defaults to 50
(FSO standard value).}

\item{share_born_female}{numeric, fraction of female babies. Defaults to
100 / 205 (FSO standard value).}

\item{n}{number of people per demographic group and year; should be
the year before \code{year_first}. Typically extracted from data frame
created with \code{propop::get_population()}.}

\item{subregional}{boolean, TRUE indicates that subregional migration
patterns (e.g., movement between municipalities within a canton)
are part of the projection.}
}
\value{
Returns an unformatted and unlabeled data frame. It includes the
number of people for each demographic group per year (starting year and
projected years. The number of rows corresponds to the product of years
and demographic groups (e.g., nationality (2) X sex (2) X age groups (101) = 404).
Variables included in the output:
\item{n}{number of people per demographic group.}
\item{IMM_INT}{number of immigrants from other countries.}
\item{MIG_CH}{number of people migrating from / to other superordinate
spatial units (typically cantons).}
\item{MIG_SUB}{number of migrants within the superordinate spatial unit
(typically a canton).}
\item{MOR}{number of deaths (among people older than 0).}
\item{EMI_INT}{number of emigrants to other countries.}
\item{ACQ}{number of foreigners who acquire Swiss citizenship
(naturalisations).}
\item{BIRTHS}{number of births.}
}
\description{
Core function that uses the cohort component method and matrix
algebra to project population development. The function can be used for
different spatial levels (e.g., cantons, municipalities) and for one scenario
at a time.

This basic function provides projections in a \bold{raw} version in which
key information is missing (e.g., which age groups the rows represent).
To conveniently obtain an enriched, more informative output,
use the \bold{wrapper function} \code{propop::propop()} (which internally uses
\code{propop::project_raw()}).

The parameters and start populations for different spatial levels
can be obtained from the Swiss Federal Statistical Office (FSO).
For instructions on how to download this information from
\href{https://www.bfs.admin.ch/bfs/en/home/services/recherche/stat-tab-online-data-search.html}{STAT-TAB},
see
\code{vignette("prepare_data", package = "propop")}.

The projection parameters need to be passed on as a single data frame to
\code{project_raw} with (with the parameters as columns). The column types, names,
and factor levels need to match those specified below.
}
\examples{
# load package data
data(fso_parameters)
data(fso_population)

# run projection
project_raw(
  parameters = fso_parameters,
  year_first = 2019,
  year_last = 2019,
  n = fso_population |> dplyr::pull(n),
  subregional = FALSE
) |>
  head(10)
}
\seealso{
\code{\link[=propop]{propop()}}
}
