% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propop-package.R
\docType{package}
\name{propop-package}
\alias{propop-package}
\title{propop: Project Population Growth in Switzerland using the Cohort Component Method}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Project the development of the population at different spatial levels (e.g., cantons, districts, municipalities) using the cohort component method, broken down by 101 one-year age classes, gender, and nationality. Includes helper functions to download required population data and projection parameters as open data for Switzerland, plus tools for model evaluation against benchmarks such as population records or a base model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://statistik-aargau.github.io/propop}
  \item \url{https://statistik-aargau.github.io/propop/}
  \item Report bugs at \url{https://github.com/statistik-aargau/propop}
}

}
\author{
\strong{Maintainer}: Norah Efosa \email{norah.efosa@ag.ch}

Authors:
\itemize{
  \item Adrian Gadient \email{adrian.gadient@ag.ch} (\href{https://orcid.org/0000-0003-0026-3153}{ORCID})
  \item Tina Cornioley \email{tina.cornioley@vd.ch}
  \item Jan Wunder \email{jan.wunder@sg.ch}
  \item Niklas Haffert \email{niklas.haffert@sg.ch}
}

Other contributors:
\itemize{
  \item Andrea Plüss \email{andrea.pluess@ag.ch} [contributor]
  \item Nadine Herrmann [contributor]
  \item Lutz Benson [contributor]
  \item Eric Meyer [contributor]
  \item Thomas Knecht [contributor]
  \item Moana Stadelmann [contributor]
  \item Statistik Aargau [funder, copyright holder]
}

}
\keyword{internal}
