% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propop.R
\name{propop}
\alias{propop}
\title{Project population development}
\usage{
propop(
  parameters,
  population,
  year_first,
  year_last,
  scenarios = NULL,
  age_groups = 101,
  fert_first = 16,
  fert_last = 50,
  share_born_female = 100/205,
  subregional = FALSE,
  binational = TRUE,
  spatial_unit = "spatial_unit"
)
}
\arguments{
\item{parameters}{data frame containing the FSO rates and numbers to run the
projection for a specific spatial level (e.g., canton, municipality).
\itemize{
\item \code{year}, character, projection year.
\item \code{spatial_unit}, character, ID of spatial entity (e.g., canton,
municipality) for which to run the projections.
\item \code{scen}, character, one or several projection scenario(s). The main
scenarios are usually "reference", "low" growth, and "high" growth.
\item \code{nat}, character, nationality (\code{ch} = Swiss; \code{int} =
foreign / international).
Required if binational = \code{TRUE}.
\item \code{sex}, character (\code{f} = female, \code{m} = male).
\item \code{age}, numeric, typically ranging from 0 to 100 (incl. >100).
\item \code{birthrate}, numeric, number of births per mother
\item \code{int_mothers}, numeric, proportion of children with
Swiss nationality born to non-Swiss mothers.
Required if binational = \code{TRUE}.
\item \code{mor}, numeric, prospective mortality rate (probability of death).
\item \code{acq}, numeric, rate of acquisition of Swiss citizenship.
Required if binational = \code{TRUE}.
\item \code{emi_int}, numeric, rate of people emigrating abroad
(number of immigrants - number of emigrants).
\item \code{emi_nat}, rate of people emigrating to other cantons.
\item \code{imm_int_n}, numeric, number of people immigrating from abroad.
\item \code{imm_nat_n}, numeric, number of people immigrating from other cantons.
\item \code{mig_sub} \bold{(optional)}, numeric, net migration per subregion; this
is the migration from / to other subregions (e.g., municipalities,
districts) within the main superordinate projection unit (e.g., a canton).
Accounts for movements between different subregions. Needs to be provided
by the user.
}}

\item{population}{data frame including the starting population of each
demographic group and each spatial unit. Possible values are the same as in
\code{parameters} (apart from year). The data frame only includes one year, usually
the one preceding the first year of the projection.
\itemize{
\item \code{year} character, should be \code{year_first} - 1.
\item \code{spatial_unit} character.
\item \code{nat} character.
\item \code{sex} character.
\item \code{age} numeric.
\item \code{n} numeric, number of people per demographic group.
}}

\item{year_first}{numeric, first year to be projected.}

\item{year_last}{numeric, last year to be projected.}

\item{scenarios}{\bold{(optional)}, character, indicating which
projection scenario(s) shall be run; the corresponding information
must be available in \code{parameters}.
Defaults to the values in variable \code{scen} in \code{parameters}.}

\item{age_groups}{numeric, number of age classes. Creates a vector with
1-year age classes running from \code{0} to (\code{age_groups} - 1). Must
currently be set to \verb{= 101} (FSO standard number of age groups).}

\item{fert_first}{numeric, first year of female fertility. Defaults to 16
(FSO standard value).}

\item{fert_last}{numeric, last year of female fertility. Defaults to 50
(FSO standard value).}

\item{share_born_female}{numeric, fraction of female babies. Defaults to
100 / 205 (FSO standard value).}

\item{subregional}{character or FALSE, indicates if subregional migration
patterns (e.g., movement between municipalities within a canton) are
part of the projection (default \code{subregional = FALSE}). Requires input
on the level of subregions (in \code{parameters} and \code{population}).
Two calculation methods are supported to distribute people between
subregions: With \code{subregional = "net"}, the net migration between
subregions is added to the population balance. Net migration numbers
must be specified in a data column \code{mig_sub} in \code{parameters}.
With \code{subregional = "rate"}, the numbers for subregional emigrants are
subtracted from the population balance, then redistributed back to all
subregional units as subregional immigration; \code{parameters} must contain
the columns \code{emi_sub} and \code{imm_sub}.}

\item{binational}{boolean, \code{TRUE} indicates that projections discriminate
between two groups of nationalities. \code{FALSE} indicates that the
projection is run without distinguishing between nationalities.}

\item{spatial_unit}{character, name of variable containing the names of the
region or subregions for which the projection shall be performed.}
}
\value{
Returns a \bold{data frame} that includes the number of people for each
demographic group per year (for projected years) and spatial unit.
The number of rows is the product of all scenarios times all years
times all demographic groups times all spatial units.
The output includes several \bold{identifiers} that indicate to which
scenario, demographic group, year, and spatial unit the results in the
rows refer to:
\item{year}{integer, indicating the projected years.}
\item{scen}{character, indicating the projected scenario(s).}
\item{spatial_unit}{factor, spatial units for which the projection
was run (e.g., canton, districts, municipalities).}
\item{age}{integer, ranging from \code{0}n to \code{100} years (including those
older than 100).}
\item{sex}{factor, female (\code{f}) and male (\code{m}).}
\item{nat}{factor, indicates if the nationality is Swiss (\code{ch}) or
international / foreign (\code{int}). This variable is only returned if
\code{binational = TRUE}.}
The output also includes columns related to the \bold{size and change
of the population:}
\item{n_jan}{numric, start-of-year population per demographic group.}
\item{n_dec}{numeric, end-of-year population per demographic group.}
\item{delta_n}{numeric, population change per demographic group from
the start to the end of the year in \emph{absolute numbers}.}
\item{delta_perc}{numeric, population change per demographic group from
the start to the end of the year in \emph{percentages}.}
The \bold{components} that are used to project the development of the
population are also included in the output:
\item{births}{numeric, number of births (non-zero values are only
available for age = 0).}
\item{mor_n}{numeric, number of deaths.}
\item{emi_int_n}{numeric, number of people who emigrate
to other countries.}
\item{emi_nat_n}{numeric, number of people who emigrate
to other cantons.}
\item{imm_int_n}{numeric, number of people who immigrate
from other countries.}
\item{imm_nat_n}{numeric, number of people who immigrate
from other cantons.}
\item{acq_n}{numeric, number of people who acquire Swiss citizenship
(only returned if  \code{binational = TRUE}.)}
}
\description{
Wrapper function to project population development using the
cohort component method. This function iterates across years and spatial units
calling the function \code{project_population.R}, which performs the calculations.

You can either use your own parameters and starting population or download
these data from the Swiss Federal Statistical Office (FSO). For instructions
on how to download this information from
\href{https://www.bfs.admin.ch/bfs/en/home/services/recherche/stat-tab-online-data-search.html}{STAT-TAB},
see \code{vignette("prepare_data", package = "propop")}.

The projection parameters need to be passed to \code{propop::propop()} as a
\bold{single data frame} (with the parameters as columns). The column types,
names, and factor levels need to match the specifications listed below under
\code{parameters}.

If nothing else is indicated in argument \code{scenarios}, \code{propop()} runs and
returns all \strong{scenarios} provided via \code{parameters}.

For more details on how to use this function to project the population
development on the level of a canton, see
\code{vignette("project_single_region", package = "propop")}.
}
\examples{
# Run projection for the sample data (whole canton of Aargau)
propop(
  parameters = fso_parameters,
  year_first = 2025,
  year_last = 2027,
  population = fso_population,
  subregional = FALSE,
  binational = TRUE
)
propop(
  parameters = fso_parameters |>
    dplyr::filter(scen == "reference" | scen == "high"),
  year_first = 2025,
  year_last = 2026,
  scenarios = c("reference", "high"),
  population = fso_population,
  subregional = FALSE,
  binational = TRUE
)
}
