% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propop_tables.R
\name{propop_tables}
\alias{propop_tables}
\title{Project population development}
\usage{
propop_tables(
  parameters,
  population,
  year_first,
  year_last,
  age_groups = 101,
  fert_first = 16,
  fert_last = 50,
  share_born_female = 100/205,
  subregional = FALSE,
  binational = TRUE,
  spatial_unit = "spatial_unit"
)
}
\description{
Wrapper function to project population development using the cohort
component method (see e.g., \href{https://www.ag.ch/media/kanton-aargau/dfr/dokumente/statistik/statistische-daten/oeffentliche-statistik/01-bevoelkerung/kantonsdaten/bevoelkerungsprognosen/bevoelkerungsprojektionen-2020-technischer-begleitbericht.pdf}{here}
for more details).
}
