% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stattab_103_snap.R
\docType{data}
\name{stattab_103_snap}
\alias{stattab_103_snap}
\title{Snapshot of STAT-TAB px 103}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 177 rows and 6 columns.
}
\source{
Data obtained from the Swiss Federal Statistical Office (FSO):
\itemize{
\item \url{https://www.pxweb.bfs.admin.ch/pxweb/en/px-x-0104020000_106/-/px-x-0104020000_106.px/}
}
}
\usage{
stattab_103_snap
}
\description{
Used for tests and to make it easier for users to look up the spelling used
on STAT-TAB.
}
\keyword{internal}
