% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stattab_pop_snap.R
\docType{data}
\name{stattab_pop_snap}
\alias{stattab_pop_snap}
\title{Snapshot of STAT-TAB that includes the population records}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2427 rows and 7 columns.
}
\source{
Federal Statistical Office:
\url{https://www.pxweb.bfs.admin.ch/pxweb/en/px-x-0102010000_101/-/px-x-0102010000_101.px/}
}
\usage{
stattab_pop_snap
}
\description{
Used for tests and to make it easier for users to look up the spelling used
on STAT-TAB.
}
\keyword{internal}
