% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_emi_mean.R
\name{calculate_emi_mean}
\alias{calculate_emi_mean}
\title{Calculate average past emigration to spatial units within the region}
\usage{
calculate_emi_mean(
  hist_data,
  n_jan,
  births,
  emi_n,
  spatial_unit,
  year_n = NULL
)
}
\arguments{
\item{hist_data}{data frame, containing aggregated migration records.
The data frame must include one row for each combination of year,
spatial unit, nationality, sex, and age.}

\item{n_jan}{character, column containing the starting population (typically
the number of people per group in January).}

\item{births}{character, column containing the number of births in the group
of 0-year olds.}

\item{emi_n}{character, column containing the total number of people emigrating per
spatial unit and demographic group.}

\item{spatial_unit}{character, column containing the spatial units.}

\item{year_n}{numeric, number of years to be considered for computing the mean.}
}
\value{
A data frame that includes the average emigration rate per demographic group
and spatial unit. \code{emi_rate} can be used as \code{emi_sub} parameter when
\code{propop::propop()} uses \code{subregional = "rate"}.
}
\description{
Calculates the mean of past emigration to other places
within the region for all spatial units and for all combinations of
age group, sex, and nationality.
}
\examples{
# Compute mean emigration rate
calculate_emi_mean(
hist_data = ag_migration_subregional,
n_jan = n_jan,
births = births,
emi_n = emi_n,
spatial_unit = spatial_unit,
year_n = NULL
)
}
\seealso{
\code{\link[=propop]{propop()}} for details on how to account for subregional migration using the rate method,
\code{\link[=calculate_shares]{calculate_shares()}} for calculating the associated immigration share \code{imm_sub}.
}
