% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_projection.R
\name{calculate_projection}
\alias{calculate_projection}
\title{Calculates the projection for ages 1-100.}
\usage{
calculate_projection(.data, subregional = FALSE)
}
\arguments{
\item{.data}{data frame, population and parameters for one year.}
}
\value{
data frame, contains the staring population \code{n_jan}, components, and
the resulting projected population for the next year \code{n_dec}.
}
\description{
Calculates the projection for ages 1-100.
}
