% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_shares.R
\name{calculate_shares}
\alias{calculate_shares}
\title{Calculate shares for distributing migration parameters to spatial units}
\usage{
calculate_shares(data_migration, column_migration)
}
\arguments{
\item{data_migration}{data frame, mutation data e.g. for international immigration or
intercantonal immigration aggregated across demographic groups.}

\item{column_migration}{string, name of the data column which contains the
data for historical migration occurrences.}
}
\description{
Calculate shares for distributing migration parameters to spatial units
}
