% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_shares.R
\name{calculate_shares}
\alias{calculate_shares}
\title{Calculate shares for distributing people among subregions}
\usage{
calculate_shares(data, col, age_group = "default")
}
\arguments{
\item{data}{data frame, historical records (e.g., immigration from
other cantons or countries) aggregated across demographic groups.}

\item{col}{character, name of the column which contains the
data for historical occurrences.}

\item{age_group}{character \bold{(optional)}, either 1-year, 5-year, or
10-year age group used as basis for calculating shares. If the argument is not
specified, the default attempts to avoid age groups without any observations.
It prioritizes age groups based on their resolution (1-year age groups =
most informative and highest priority, 10-year age groups = least informative
and lowest priority).
Users can override the default and enforce the use of a specific age group
for all demographic groups by setting the argument to "age_group_5" or
"age_group_10".}
}
\value{
Returns the input data frame with the following new columns:
\itemize{
\item \code{age_group_5}: character, name of the 5-year age group to which the 1-year
age group is assigned to.
\item \code{sum_5}: numeric, total number of people in the 5-year age group.
\item \code{prop_5}: numeric, proportion of the the 5-year age group total that
is allocated to each 1-year age group.
\item \code{age_group_10}: character, name of the 10-year age group to which the 1-year
age group is assigned to.
\item \code{sum_10}: numeric, total number of people in the 10-year age group.
\item \code{prop_10}: numeric, proportion of the the 10-year age group total that
is allocated to each 1-year age group.
\item \code{use_age_group}: character, preference for 1-year, 5-year or 10-year age
group. Defaults to \code{age_group_1} if at least one observation was recorded.
\item \code{use_share}: numeric, share to be used according to \code{use_age_group}.
\item \code{use_share_sum}: numeric, total per demographic group and across all
spatial units.
\item \code{share}: numeric, the spatial unit's share relative to the total (across all
spatial units) of people within the same demographic group.
}
}
\description{
Calculate shares for distributing people among subregions
}
