% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_population.R
\name{get_population}
\alias{get_population}
\title{Get population data from FSO}
\source{
Federal Statistical Office:
\url{https://www.pxweb.bfs.admin.ch/pxweb/en/px-x-0102010000_101/-/px-x-0102010000_101.px/}
}
\usage{
get_population(
  number_fso = "px-x-0102010000_101",
  year,
  year_last = NULL,
  spatial_units
)
}
\arguments{
\item{number_fso}{character, px-x table ID for population records,
defaults to \verb{px-x-0102010000_101}.}

\item{year}{numeric, year for which the population records are to
be downloaded. This usually is the starting population. To download
longer time periods, use \code{year} to indicate the onset of the period.}

\item{year_last}{numeric \bold{(optional)}; specifies the final year of the
time period for which data will be downloaded.}

\item{spatial_units}{character vector, indicating at least one spatial
entity for which the projection will be run. Typically a canton, districts,
or municipalities.}
}
\value{
A data frame. For each of the four demographic groups (female / male,
Swiss / foreign nationals), there are 101 age classes, resulting in a total
of 404 rows per requested year and spatial unit. Columns included in the
returned data frame:
\describe{
\item{year}{numeric, year in which the population was recorded.}
\item{spatial_unit}{character, indicating the spatial entities (e.g.,
cantons, districts, municipalities).}
\item{nat}{character, ch = Swiss, int = foreign / international.}
\item{sex}{character f = female, m = male.}
\item{age}{numeric, 101 one-year age classes, ranging from 0 to 100
(including those older than 100).}
\item{n}{numeric, number of people per year, spatial entity, and
demographic group.}
}
}
\description{
Users who do not have the required population data can use this convenience
function to get the mandatory starting population for
\code{propop::propop()} from the Federal Statistical Office (FSO). The function
can also be used to obtain the population records for several years
(e.g., for model performance evaluations). This function can be used to
obtain data at various spatial levels (e.g., cantons, municipalities).
The most recent data are usually about 6 to 18 months old.

To get the population data, you must use the \strong{spelling} defined in the
corresponding FSO table (STATTAB cube
\href{https://www.bfs.admin.ch/asset/de/32207872}{px-x-0102010000_101}).
Inspecting the column 'valueTexts' in the following package data may also
help: \code{data('stattab_pop_snap')}.

Changes to the API interface may break this function.
}
\examples{
\dontshow{if ((Sys.getenv("RUN_EXPENSIVE_TESTS") == "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
get_population(
  number_fso = "px-x-0102010000_101",
  year = 2020,
  year_last = 2023,
  spatial_units = "- Aargau"
)
get_population(
  year = 2023,
  spatial_units = c("- Aargau", "......0301 Aarberg")
)
}
\dontshow{\}) # examplesIf}
}
