% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_evaluation.R
\name{prepare_evaluation}
\alias{prepare_evaluation}
\title{Prepare data for evaluation}
\usage{
prepare_evaluation(
  data_benchmark,
  n_benchmark,
  data_projected,
  n_projected,
  age_groups = NULL
)
}
\arguments{
\item{data_benchmark}{data frame containing benchmark data (e.g., actual /
official population records; can be obtained with \code{propop::get_population()}).}

\item{n_benchmark}{numeric column containing the benchmark population of each
demographic group.}

\item{data_projected}{data frame containing population projections; can be
created with \code{propop::propop()}.}

\item{n_projected}{numeric column containing the projected size of each
demographic group.}

\item{age_groups}{character, optional argument with options \code{"age_groups_3"}
or \code{"age_groups_5"}.
The option \code{"age_groups_3"} groups the data into
three age ranges (0-19, 20-64, 65 years and older).
The option \code{"age_groups_5"} groups the data into
five age ranges (0-19, 20-39, 40-59, 60-79, 80 years and older).
Using aggregated groups will lead to smaller
projection errors than using 101 age classes.
Defaults to using 101 one-year age classes if no option is chosen.}
}
\value{
Returns a data frame with the number of people from the benchmark and
from the projection. Each row contains a unique combination of year, spatial
unit, and demographic group.
}
\description{
Takes population projections and benchmark data (e.g., population records)
and prepares a combined data frame to evaluate the performance of the
projection. For more details on usage, see
\code{vignette("evaluate", package = "propop")}.
}
\section{Input data and variables}{

Both input data frames must contain the following variables for the
\bold{same range of years}:
\describe{
\item{year}{character, year in which the population was recorded.}
\item{spatial_unit}{character, indicating the projected spatial entities
(e.g., cantons, districts, municipalities).}
\item{nat}{character, \code{ch} = Swiss, \code{int} = foreign / international.}
\item{sex}{character, \code{f} = female, \code{m} = male.}
\item{age}{numeric, 101 one-year age classes, ranging from 0 to 100
(including those older than 100).}
\item{n}{numeric, number of people per year, spatial entity, and
demographic group.}
}
}

\examples{
\dontrun{
combined <- prepare_evaluation(
  data_benchmark = output_get_population,
  data_projected = output_propop
)
combined_grouped <- prepare_evaluation(
  data_benchmark = output_get_population,
  data_projected = output_propop,
  age_groups = "age_groups_3"
)
}
}
