% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_raw.R
\name{project_raw}
\alias{project_raw}
\title{Project population development (raw results)}
\usage{
project_raw(
  parameters,
  year_first,
  year_last,
  age_groups = 101,
  fert_first = 16,
  fert_last = 50,
  share_born_female = 100/205,
  n,
  subregional
)
}
\arguments{
\item{parameters}{data frame containing the FSO rates and numbers to run the
projection for a specific spatial level (e.g., canton, municipality).
\itemize{
\item \code{year}: projection year.
\item \code{spatial_unit}: ID of spatial entity (e.g., canton, municipality) for
which to run the projections.
\item \code{scen}: projection scenario, used to subset data frames with multiple
scenarios (r = reference, l = low growth, h = high growth scenario).
\item \code{nat}: nationality (ch = Swiss; int = foreign / international).
\item \code{sex}: sex (f = female, m = male).
\item \code{age}: age classes; typically ranging from 0 to 100 (incl. >100).
\item \code{birthrate}: numeric, total number of live human births per 1,000
inhabitants.
\item \code{int_mothers} proportion of children with Swiss nationality born to
non-Swiss mothers.
\item \code{mor}: prospective mortality rate (probability of death).
\item \code{acq}: rate of acquisition of Swiss citizenship.
\item \code{emi_int}: rate of people emigrating abroad.
\item \code{emi_nat}: rate of people emigrating to other cantons.
\item \code{imm_int_n}: number of people immigrating from abroad.
\item \code{imm_nat_n}: number of people immigrating from other cantons.
\item \code{mig_sub}: within canton net migration. Useful to account for movements
between different subregions (e.g., municipalities). This argument is
\bold{optional.}
}}

\item{year_first}{numeric, first year to be projected.}

\item{year_last}{numeric, last year to be projected.}

\item{age_groups}{numeric, number of age classes. Creates a vector with
1-year age classes running from \code{0} to (\code{age_groups} - 1). Defaults to
\code{101} (FSO standard number of age groups).}

\item{fert_first}{numeric, first year of female fertility. Defaults to 16
(FSO standard value).}

\item{fert_last}{numeric, last year of female fertility. Defaults to 50
(FSO standard value).}

\item{share_born_female}{numeric, fraction of female babies. Defaults to
100 / 205 (FSO standard value).}

\item{n}{number of people per demographic group and year; should be
the year before \code{year_first}. Typically extracted from data frame
created with \code{propop::get_population()}.}

\item{subregional}{boolean, TRUE indicates that subregional migration
patterns (e.g., movement between municipalities within a canton)
are part of the projection.}
}
\value{
Returns an unformatted and unlabeled data frame. It includes the
number of people for each demographic group per year (starting year and
projected years. The number of rows corresponds to the product of years
and demographic groups (e.g., nationality (2) X sex (2) X age groups (101) = 404).
Variables included in the output:
\item{n}{number of people per demographic group.}
\item{IMM_INT}{number of immigrants from other countries.}
\item{MIG_NAT}{number of people migrating from / to other superordinate
spatial units (typically cantons).}
\item{MIG_SUB}{number of migrants within the superordinate spatial unit
(typically a canton).}
\item{MOR}{number of deaths (among people older than 0).}
\item{EMI_INT}{number of emigrants to other countries.}
\item{ACQ}{number of foreigners who acquire Swiss citizenship
(naturalisations).}
\item{BIRTHS}{number of births.}
}
\description{
Core function that uses the cohort component method and matrix
algebra to project population development. The function can be used for
different spatial levels (e.g., cantons, municipalities) and for one scenario
at a time.

This function provides projections in a \bold{raw} version in which
key information is missing (e.g., which age groups the rows represent).
To conveniently obtain an enriched, more informative output,
use the \bold{wrapper function} \code{propop::propop()} (which internally uses
\code{propop::project_raw()}).

The parameters and starting populations for different spatial levels
can be obtained from the Swiss Federal Statistical Office (FSO).
For instructions on how to download this information from
\href{https://www.bfs.admin.ch/bfs/en/home/services/recherche/stat-tab-online-data-search.html}{STAT-TAB},
see
\code{vignette("prepare_data", package = "propop")}.

The projection parameters need to be passed on as a single data frame to
\code{project_raw} with (with the parameters as columns). The column types, names,
and factor levels need to match those specified below.

The method used to calculate the projections is a 'cohort-component
analysis' implemented with matrices due to programming performance benefit
compared to data frames. In a nutshell, the starting population ('n') is
multiplied by the survival rate to obtain the number of people which
transition into the projected next year (year + 1). Then, the absolute
number of people immigrating from outside Switzerland and the migration saldo
for people from outside the respective canton is added to the surviving
population. This results in the starting population for projection the next
year. Newborn children are added aeparately to the new starting population
of each year.

The starting population is clustered in 404 groups: 101 age groups times
two nationalities times 2 genders. The survival rate is calculated in the
function 'create_transition_matrix()' resulting in the matrix 'L'. We use the
rates for mortality, emigration towards countries outside Switzerland and the
rate for the acquisition of the Swiss citizenship by the foreign population
to calculate survival rates. The model from the FSO also includes the rate of
emigration to other cantons in the survival rate. In contrast, we include the
immi- and emigration from and to other cantons by adding the migration
balance (German = 'saldo') (immigration + emigration) afterwards.

Steps in this function:
\enumerate{
\item Checks: Checking input data and parameter settings for correct formats.
\item Data preparation: Preparing vectors e.g. for the projection time frame and
creating empty vectors to be filled with data later on.
\item Loop over years for calculating the projections
\itemize{
\item Subsetting parameters: Depending on the selected projection year and on
the demographic unit, the parameters for mortality, fertility, acquisition
of the Swiss citizenship as well as migration parameters are subset by
demographic group.
\item Create matrices: Matrices are build for the survival rate, mortality,
fertility and for calculating the number of newborn babies.
\item Creating vectors: Vectors are built for mortality and migration parameters.
\item Projection: The transition matrix 'L' is multiplied by the starting
population for the next year. Migrating people are added in absolute
numbers. People that are 100 years old and older are clustered into one
age group (age = 100). The newborn babies are added to the resulting
starting population for the next projection year.
}
\item Aggregating the data: All projected years are aggregated into one data
frame. The function 'propop()', in which this function is contained,
automatically adds relevant meta data to the results.
}
}
\examples{
# load package data
data(fso_parameters)
data(fso_population)

# run projection
project_raw(
  parameters = fso_parameters,
  year_first = 2019,
  year_last = 2019,
  n = fso_population |> dplyr::pull(n),
  subregional = FALSE
) |>
  head(10)
}
\seealso{
\code{\link[=propop]{propop()}}
}
