% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propop.R
\name{propop}
\alias{propop}
\title{Project population development (enriched results)}
\usage{
propop(
  parameters,
  population,
  year_first,
  year_last,
  age_groups = 101,
  fert_first = 16,
  fert_last = 50,
  share_born_female = 100/205,
  subregional = FALSE,
  binational = TRUE,
  spatial_unit = "spatial_unit"
)
}
\arguments{
\item{parameters}{data frame containing the FSO rates and numbers to run the
projection for a specific spatial level (e.g., canton, municipality).
\itemize{
\item \code{year}, character, projection year.
\item \code{spatial_unit}, character, ID of spatial entity (e.g., canton,
municipality) for which to run the projections.
\item \code{scen}, character, projection scenario, is used to subset data frames
with multiple scenarios (r = reference, l = low growth, h = high growth).
\item \code{nat} \bold{(optional)}, character, nationality (ch = Swiss; int =
foreign / international).
\item \code{sex}, character (f = female, m = male).
\item \code{age}, numeric, typically ranging from 0 to 100 (incl. >100).
\item \code{birthrate}, numeric, number of children per year.
\item \code{int_mothers} \bold{(optional)}, numeric, proportion of children with
Swiss nationality born to non-Swiss mothers.
\item \code{mor}, numeric, prospective mortality rate (probability of death).
\item \code{acq} \bold{(optional)}, numeric, rate of acquisition of Swiss citizenship.
\item \code{emi_int}, numeric, rate of people emigrating abroad.
(number of immigrants - number of emigrants).
\item \code{emi_nat}: rate of people emigrating to other cantons.
\item \code{imm_int_n}, numeric, number of people immigrating from abroad.
\item \code{imm_nat_n}: number of people immigrating from other cantons.
\item \code{mig_sub} \bold{(optional)}, numeric, within canton net migration. Useful
to account for movements between different subregions (e.g., municipalities).
}}

\item{population}{data frame including the starting population of each
demographic group. Possible values are the same as in \code{parameters} (apart
from year). The data frame only includes one year, usually the one preceding
the first projected year.
\itemize{
\item \code{year} character, should be \code{year_first} - 1.
\item \code{spatial_unit} character.
\item \code{nat} character.
\item \code{sex} character.
\item \code{age} numeric.
\item \code{n} numeric, number of people per demographic group.
}}

\item{year_first}{numeric, first year to be projected.}

\item{year_last}{numeric, last year to be projected.}

\item{age_groups}{numeric, number of age classes. Creates a vector with
1-year age classes running from \code{0} to (\code{age_groups} - 1). Defaults to
\code{101} (FSO standard number of age groups).}

\item{fert_first}{numeric, first year of female fertility. Defaults to 16
(FSO standard value).}

\item{fert_last}{numeric, last year of female fertility. Defaults to 50
(FSO standard value).}

\item{share_born_female}{numeric, fraction of female babies. Defaults to
100 / 205 (FSO standard value).}

\item{subregional}{boolean, \code{TRUE} indicates that subregional migration
patterns (e.g., movement between municipalities within a canton)
are part of the projection. Requires input (parameters and population)
on the level of subregions.}

\item{binational}{boolean, \code{TRUE} indicates that projections discriminate
between two groups of nationalities. \code{FALSE} indicates that only one
projection is run without distinguishing between nationalities.}

\item{spatial_unit}{character, name of variable containing the names of the
region or subregions for which the projection shall be performed.}
}
\value{
Returns a data frame that includes the number of people for each demographic
group per year (for the starting year and each projected year). The
number of rows is the product of all years times all demographic groups.
The output includes several \bold{identifiers} that indicate to which
demographic group and year the results in the rows refer to.
\item{year}{integer, indicating starting year / projected years.}
\item{spatial_unit}{factor, spatial units for which the projection
was run (e.g., canton, municipalities, districts).}
\item{age}{integer.}
\item{sex}{factor, female (f) and male (m).}
\item{nat}{factor, indicates if the nationality is Swiss (ch) or
international / foreign (int). This variable is only returned if
\code{binational = TRUE}.}
The output also includes columns related to the \bold{size and change
of the population:}
\item{n}{numric, end-of-year population per demographic group.}
\item{n_1}{numeric, number of people of the particular
demographic group by the end of the following year.}
\item{delta_n}{numeric, population change per demographic group from
current to next year in absolute numbers.}
\item{delta_perc}{numeric, population change per demographic group from
current to next year in percentages.}
The \bold{components} that are used to project the development of the population
are also included in the output:
\item{births}{numeric, number of births (values only available for
age = 0).}
\item{mor}{numeric, number of deaths.}
\item{emi_int}{numeric, number of people who emigrate
to other countries.}
\item{emin_nat}{numeric, number of people who emigrate
to other cantons.}
\item{imm_int}{numeric, number of people who immigrate
from other countries.}
\item{imm_nat}{numeric, number of people who immigrate
from other cantons.}
\item{acq}{numeric, number of people who acquire Swiss citizenship.}
}
\description{
Wrapper function to project population development using the cohort
component method (see e.g., \href{https://www.ag.ch/media/kanton-aargau/dfr/dokumente/statistik/statistische-daten/oeffentliche-statistik/01-bevoelkerung/kantonsdaten/bevoelkerungsprognosen/bevoelkerungsprojektionen-2020-technischer-begleitbericht.pdf}{here}
for more details).

You can either use your own parameters and starting population or download these
data from the Swiss Federal Statistical Office (FSO). For instructions on how
to download this information from
\href{https://www.bfs.admin.ch/bfs/en/home/services/recherche/stat-tab-online-data-search.html}{STAT-TAB},
see \code{vignette("prepare_data", package = "propop")}.

For more details on how to use this function to project the population
development on the level of a canton, see
\code{vignette("run_projections", package = "propop")}.

The projection parameters need to be passed to \code{propop::propop()} as a
\bold{single data frame} (with the parameters as columns). The column types,
names, and factor levels need to match the specifications listed below under
\code{parameters}:
}
\examples{
# Run projection for the sample data (whole canton of Aargau)
propop(
  parameters = fso_parameters,
  year_first = 2019,
  year_last = 2022,
  population = fso_population,
  subregional = FALSE,
  binational = TRUE
)
}
