% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propop.R
\name{propop}
\alias{propop}
\title{Project population development (enriched results)}
\usage{
propop(
  parameters,
  population,
  year_first,
  year_last,
  age_groups = 101,
  fert_first = 16,
  fert_last = 50,
  share_born_female = 100/205,
  subregional,
  binational,
  spatial_unit = "spatial_unit"
)
}
\arguments{
\item{parameters}{data frame containing the FSO rates and numbers to run the
projection for a specific spatial level (e.g., canton, municipality).
\itemize{
\item \code{year} character, projection year.
\item \code{spatial_unit}, character, ID of spatial entity (e.g., canton,
municipality) for which to run the projections.
\item \code{scen}, character, projection scenario, is used to subset data frames
with multiple scenarios (r = reference, l = low growth, h = high growth).
\item \code{nat}, character, OPTIONAL; nationality (ch = Swiss; int = foreign/international).
\item \code{sex}, character (f = female, m = male).
\item \code{age}, numeric, typically ranging from 0 to 100 (incl. >100).
\item \code{birth_rate}, numeric, number of children per year.
\item \code{births_int_ch}, numeric, OPTIONAL; proportion of children with Swiss nationality
born to non-Swiss mothers.
\item \code{mor}, numeric, prospective mortality rate (probability of death).
\item \code{acq}, numeric, OPTIONAL; rate of acquisition of Swiss citizenship.
\item \code{emi}, numeric, rate of people emigrating abroad.
\item \code{mig_ch}, numeric, national / inter-cantonal net migration
(number of immigrants - number of emigrants).
\item \code{imm_int}, numeric, number of people immigrating from abroad.
\item \code{mig_sub}, numeric, within canton net migration. Useful to account
for movements between different subregions (e.g., municipalities).
This argument is \bold{optional.}
}}

\item{population}{data frame including the starting population of each
demographic group. Possible values are the same as in \code{parameters} (apart
from year). The data frame only includes one year, usually the one preceding
the first projected year.
\itemize{
\item \code{year} character, should be \code{year_first} - 1.
\item \code{spatial_unit} character.
\item \code{nat} character.
\item \code{sex} character.
\item \code{age} numeric.
\item \code{n} numeric, number of people per demographic group.
}}

\item{year_first}{numeric, first year to be projected.}

\item{year_last}{numeric, last year to be projected.}

\item{age_groups}{numeric, number of age classes. Creates a vector with
1-year age classes running from \code{0} to (\code{age_groups} - 1). Defaults to
\code{101} (FSO standard number of age groups).}

\item{fert_first}{numeric, first year of female fertility. Defaults to 16
(FSO standard value).}

\item{fert_last}{numeric, last year of female fertility. Defaults to 50
(FSO standard value).}

\item{share_born_female}{numeric, fraction of female babies. Defaults to
100 / 205 (FSO standard value).}

\item{subregional}{boolean, TRUE indicates that subregional migration
patterns (e.g., movement between municipalities within a canton)
are part of the projection.}

\item{binational}{boolean, TRUE indicates that projections discriminate
between two groups of nationalities. FALSE indicates no distinction
between groups of nationalities.}

\item{spatial_unit}{character, name of variable containing the names of the
region or subregions for which the projection shall be performed.}
}
\value{
Returns a data frame that includes the number of people for each demographic
group per year (for the starting year and each projected year).
The number of rows is the product of all years times all demographic
groups (e.g., nationality (2) * sex (2) * age groups (101) = 404).
Variables included in the output:
\item{age}{integer.}
\item{sex}{factor, female (f) and male (m).}
\item{nat}{factor, Swiss (ch) and international / foreign (int).}
\item{year}{integer, indicating starting year / projected years.}
\item{spatial_levels}{factor, spatial levels for which the projection
was run (e.g., canton, municipalities).}
\item{n}{double, number of people per demographic group.}
}
\description{
Project population development using the cohort component method (see e.g.,
\href{https://www.ag.ch/media/kanton-aargau/dfr/dokumente/statistik/statistische-daten/oeffentliche-statistik/01-bevoelkerung/kantonsdaten/bevoelkerungsprognosen/bevoelkerungsprojektionen-2020-technischer-begleitbericht.pdf}{here}
for more details).
This \bold{wrapper function} runs the projection and returns a clean
data frame with clearly labeled variables and variable levels.

The parameters and start population can be obtained from the Swiss Federal
Statistical Office (FSO). For instructions on how to download this
information from
\href{https://www.bfs.admin.ch/bfs/en/home/services/recherche/stat-tab-online-data-search.html}{STAT-TAB},
see \code{vignette("prepare_data", package = "propop")}.

For more details on how to use this function to project the population
development on the level of a canton, see
\code{vignette("run_projections", package = "propop")}.

The projection parameters need to be passed to \code{propop::propop()} as a
\bold{single data frame} (with the parameters as columns). The column types,
names, and factor levels need to match the specifications listed below under
\code{parameters}:
}
\examples{
# Run projection for the sample data (whole canton of Aargau)
propop(
  parameters = fso_parameters,
  year_first = 2019,
  year_last = 2022,
  population = fso_population,
  subregional = FALSE,
  binational = TRUE
)
}
